Failed approach to refactoring the Bottles.verse case statement
This was an attempt to break up the case statement to abstract out the
notion of current count and next count, tolerating a little redundancy
for now in having two steps for the case analysis. Unfortunately it
didn't cover the implications on the pluralization of the nouns. To add
that coverage would increase the redundancy even further and make it
not worth it.